Fix 64-bit literals, struct array fields, and funcptr array parsing
- Support large numeric literals (>32-bit) by widening Expr.num to long
  and using strtol() instead of atoi(), with sign-extension for non-L
  suffixed values in the unsigned 32-bit range
- Fix selfhost struct array field counting (> 1 → > 0) in
  cg_field_index() and cg_struct_nfields() to handle single-element arrays
- Add array dimension parsing for function pointer globals in selfhost
  (e.g. int (*ops[3])(int, int))

+++ cc_in_c/ast.c
@@ -21,7 +21,7 @@ void stmtarray_push(StmtArray *a, Stmt *s) {
     a->data[a->len++] = s;
 }
 
-Expr *new_num(int val) {
+Expr *new_num(long val) {
     Expr *e = xmalloc(sizeof(Expr));
     e->kind = ND_NUM;
     e->u.num = val;
